Move ProcessedTx into persona for avoiding circular dependencies
