+++ packages/sanes-chrome-extension/src/routes/account/components/ListTxs.tsx
@@ -1,28 +1,12 @@
-import { Amount } from '@iov/bcp';
 import * as React from 'react';
-import { ReadonlyDate } from 'readonly-date';
 import Block from 'medulas-react-components/lib/components/Block';
 import Hairline from 'medulas-react-components/lib/components/Hairline';
 import Typography from 'medulas-react-components/lib/components/Typography';
 import { List, ListItem, ListItemText } from 'medulas-react-components/lib/components/List';
+import { ProcessedTx } from '../../../logic/persona';
 import upToDate from '../assets/uptodate.svg';
 import Tx from './Tx';
 import EmptyList from './Empty';
-
-interface TxProps {
-  readonly id: string;
-  readonly recipient: string;
-  readonly signer: string;
-  readonly amount: Amount;
-  readonly memo?: string;
-}
-
-export interface ProcessedTx extends TxProps {
-  readonly time: ReadonlyDate;
-  readonly received: boolean;
-  readonly success: boolean;
-  readonly err?: any; // eslint-disable-line
-}
 
 interface Props {
   readonly txs: ReadonlyArray<ProcessedTx>;
